+++ src/main/java/com/cloudogu/scm/review/emailnotification/EmailNotificationHook.java
@@ -25,6 +25,8 @@ package com.cloudogu.scm.review.emailnotification;
 
 import com.cloudogu.scm.review.comment.service.CommentEvent;
 import com.cloudogu.scm.review.comment.service.MentionEvent;
+import com.cloudogu.scm.review.comment.service.Reply;
+import com.cloudogu.scm.review.comment.service.ReplyEvent;
 import com.cloudogu.scm.review.pullrequest.service.BasicPullRequestEvent;
 import com.cloudogu.scm.review.pullrequest.service.PullRequest;
 import com.cloudogu.scm.review.pullrequest.service.PullRequestApprovalEvent;
@@ -33,6 +35,7 @@ import com.cloudogu.scm.review.pullrequest.service.PullRequestMergedEvent;
 import com.cloudogu.scm.review.pullrequest.service.PullRequestRejectedEvent;
 import com.github.legman.Subscribe;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.shiro.SecurityUtils;
 import sonia.scm.EagerSingleton;
 import sonia.scm.HandlerEventType;
 import sonia.scm.mail.api.MailSendBatchException;
@@ -43,6 +46,8 @@ import javax.inject.Inject;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
+
+import static java.util.stream.Collectors.toSet;
 
 @Slf4j
 @EagerSingleton
@@ -58,14 +63,38 @@ public class EmailNotificationHook {
 
   @Subscribe
   public void handlePullRequestEvents(PullRequestEvent event) {
-    handleEvent(event, new PullRequestEventMailTextResolver(event));
+    PullRequest pullRequest = event.getPullRequest();
+    EMailRecipientHelper eMailRecipientHelper = new EMailRecipientHelper(pullRequest);
+    Set<String> subscriberWithoutReviewers = eMailRecipientHelper.getSubscriberWithoutReviewers();
+    Set<String> reviewers = eMailRecipientHelper.getSubscribingReviewers();
+
+    handleEvent(event, new PullRequestEventMailTextResolver(event, false), pullRequest, subscriberWithoutReviewers);
+    handleEvent(event, new PullRequestEventMailTextResolver(event, true), pullRequest, reviewers);
   }
 
   @Subscribe
   public void handleCommentEvents(CommentEvent event) {
     if (!isSystemComment(event)) {
-      handleEvent(event, new CommentEventMailTextResolver(event));
+      PullRequest pullRequest = event.getPullRequest();
+      handleEvent(event, new CommentEventMailTextResolver(event), pullRequest, getSubscribersWithoutCurrentUser(pullRequest));
     }
+  }
+
+  @Subscribe
+  public void handleReplyEvents(ReplyEvent event) {
+    PullRequest pullRequest = event.getPullRequest();
+    Set<String> authorsInThread = getAuthorsInThread(event);
+    Set<String> subscribers = getSubscribersWithoutCurrentUser(pullRequest);
+    subscribers.removeAll(authorsInThread);
+    handleEvent(event, new CommentEventMailTextResolver(event), pullRequest, subscribers);
+    handleEvent(event, new ReplyEventMailTextResolver(event), pullRequest, authorsInThread);
+  }
+
+  private Set<String> getAuthorsInThread(ReplyEvent replyEvent) {
+    Set<String> involvedUsers = new HashSet<>();
+    involvedUsers.add(replyEvent.getRootComment().getAuthor());
+    replyEvent.getRootComment().getReplies().stream().map(Reply::getAuthor).forEach(involvedUsers::add);
+    return filterCurrentUser(involvedUsers);
   }
 
   @Subscribe
@@ -77,7 +106,9 @@ public class EmailNotificationHook {
     } else {
       newMentions = event.getItem().getMentionUserIds();
     }
-    service.sendEmails(new MentionEventMailTextResolver(event), newMentions, Collections.emptySet());
+    MailTextResolver mailTextResolver = new MentionEventMailTextResolver(event);
+    service.sendEmail(newMentions, mailTextResolver);
+    service.sendEmail(Collections.emptySet(), mailTextResolver);
   }
 
   private boolean isSystemComment(CommentEvent event) {
@@ -90,30 +121,44 @@ public class EmailNotificationHook {
 
   @Subscribe
   public void handleMergedPullRequest(PullRequestMergedEvent event) {
-    handleEvent(event, new PullRequestMergedMailTextResolver(event));
+    PullRequest pullRequest = event.getPullRequest();
+    handleEvent(event, new PullRequestMergedMailTextResolver(event), pullRequest, getSubscribersWithoutCurrentUser(pullRequest));
   }
 
   @Subscribe
   public void handleRejectedPullRequest(PullRequestRejectedEvent event) {
-    handleEvent(event, new PullRequestRejectedMailTextResolver(event));
+    PullRequest pullRequest = event.getPullRequest();
+    handleEvent(event, new PullRequestRejectedMailTextResolver(event), pullRequest, getSubscribersWithoutCurrentUser(pullRequest));
   }
 
   @Subscribe
   public void handlePullRequestApproval(PullRequestApprovalEvent event) {
-    handleEvent(event, new PullRequestApprovalMailTextResolver(event));
+    PullRequest pullRequest = event.getPullRequest();
+    handleEvent(event, new PullRequestApprovalMailTextResolver(event), pullRequest, getSubscribersWithoutCurrentUser(pullRequest));
   }
 
-  private void handleEvent(BasicPullRequestEvent event, MailTextResolver mailTextResolver) {
-    PullRequest pullRequest = event.getPullRequest();
+  private void handleEvent(BasicPullRequestEvent event, MailTextResolver mailTextResolver, PullRequest pullRequest, Set<String> recipients) {
     Repository repository = event.getRepository();
     if (pullRequest == null || repository == null) {
-      log.error("Repository or Pull Request not found in the event {}", event);
+      log.warn("Repository or Pull Request not found in the event {}", event);
       return;
     }
     try {
-      service.sendEmails(mailTextResolver, pullRequest.getSubscriber(), pullRequest.getReviewer().keySet());
+      service.sendEmail(recipients, mailTextResolver);
     } catch (Exception e) {
       log.warn("Error on sending Email", e);
     }
+  }
+
+  private Set<String> getSubscribersWithoutCurrentUser(PullRequest pullRequest) {
+    return filterCurrentUser(pullRequest.getSubscriber());
+  }
+
+  private Set<String> filterCurrentUser(Set<String> users) {
+    String currentUser = SecurityUtils.getSubject().getPrincipal().toString();
+    if (users.contains(currentUser)) {
+      return users.stream().filter(s -> !s.equals(currentUser)).collect(toSet());
+    }
+    return users;
   }
 }