+++ src/main/java/com/cloudogu/scm/review/emailnotification/MentionEventMailTextResolver.java
@@ -25,6 +25,7 @@ package com.cloudogu.scm.review.emailnotification;
 
 import com.cloudogu.scm.review.comment.service.MentionEvent;
 import lombok.extern.slf4j.Slf4j;
+import sonia.scm.mail.api.Topic;
 
 import java.util.Locale;
 import java.util.Map;
@@ -53,9 +54,14 @@ public class MentionEventMailTextResolver extends BasicPRMailTextResolver<Mentio
   }
 
   @Override
-  public Map<String, Object> getContentTemplateModel(String basePath, boolean isReviewer) {
-    Map<String, Object> model = getTemplateModel(basePath, mentionEvent, false);
+  public Map<String, Object> getContentTemplateModel(String basePath) {
+    Map<String, Object> model = getTemplateModel(basePath, mentionEvent);
     model.put("comment", mentionEvent.getItem());
     return model;
+  }
+
+  @Override
+  public Topic getTopic() {
+    return TOPIC_MENTIONS;
   }
 }