+++ src/test/java/com/cloudogu/scm/review/emailnotification/EmailNotificationServiceTest.java
@@ -24,11 +24,6 @@
 package com.cloudogu.scm.review.emailnotification;
 
 import com.google.common.collect.Sets;
-import org.apache.shiro.subject.Subject;
-import org.apache.shiro.util.ThreadContext;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Answers;
@@ -40,16 +35,13 @@ import org.mockito.quality.Strictness;
 import sonia.scm.config.ScmConfiguration;
 import sonia.scm.mail.api.MailSendBatchException;
 import sonia.scm.mail.api.MailService;
-import sonia.scm.mail.api.MailTemplateType;
 
 import java.util.Collections;
 import java.util.Map;
 
 import static java.util.Locale.ENGLISH;
 import static java.util.Locale.GERMAN;
-import static org.mockito.Mockito.anyBoolean;
 import static org.mockito.Mockito.anyString;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
@@ -76,127 +68,46 @@ class EmailNotificationServiceTest {
   @Mock
   private MailTextResolver mailTextResolver;
 
+  private final Map<String,Object> subscriberModel = Collections.singletonMap("subscriber", true);
 
   @Test
   void shouldNotSendEmailForNotConfiguredMailServer() throws Exception {
     when(mailService.isConfigured()).thenReturn(false);
 
-    service.sendEmails(mailTextResolver, Collections.singleton("trillian"), Collections.emptySet());
+    service.sendEmail(Collections.singleton("trillian"), mailTextResolver);
+    service.sendEmail(Collections.emptySet(), mailTextResolver);
 
     verify(mailService).isConfigured();
     verifyNoMoreInteractions(mailService);
   }
 
-  @Nested
-  class WithPrincipal {
+  @Test
+  void shouldSendEmailToRecipient() throws MailSendBatchException {
+    mockDependencies();
+    when(mailService.isConfigured()).thenReturn(true);
+    when(mailService.emailTemplateBuilder()).thenReturn(envelopeBuilder);
 
-    @Mock
-    private Subject subject;
+    service.sendEmail(Sets.newHashSet("dent", "trillian"), mailTextResolver);
 
-    private Map<String,Object> subscriberModel = Collections.singletonMap("subscriber", true);
-
-    private Map<String,Object> reviewerModel = Collections.singletonMap("reviewer", true);
-
-    @BeforeEach
-    void setUpSubject() {
-      ThreadContext.bind(subject);
-    }
-
-    @AfterEach
-    void tearDownSubject() {
-      ThreadContext.unbindSubject();
-    }
-
-    @Test
-    void shouldSendEmailForSubscriberOnly() throws MailSendBatchException {
-      mockDependencies("slarti");
-
-      service.sendEmails(mailTextResolver, Sets.newHashSet("dent", "trillian"), Collections.singleton("marvin"));
-
-      verifier().send("dent", "trillian").notSend("marvin");
-    }
-
-    @Test
-    void shouldNotSendEmailToPrincipal() throws MailSendBatchException {
-      mockDependencies("dent");
-
-      service.sendEmails(mailTextResolver, Sets.newHashSet("dent", "trillian"), Collections.emptySet());
-
-      verifier().send("trillian").notSend("dent");
-    }
-
-    @Test
-    void shouldSendEmailForReviewers() throws MailSendBatchException {
-      mockDependencies("marvin");
-
-      service.sendEmails(mailTextResolver, Collections.singleton("dent"), Collections.singleton("dent"));
-
-      verifier().send("dent").reviewer();
-    }
-
-    @Test
-    void shouldSendEmailForSubscriber() throws MailSendBatchException {
-      mockDependencies("marvin");
-
-      service.sendEmails(mailTextResolver, Collections.singleton("dent"), Collections.emptySet());
-
-      verifier().send("dent").subscriber();
-    }
-
-    private Verifier verifier() {
-      verify(envelopeBuilder).fromCurrentUser();
-      return new Verifier();
-    }
-
-    private void mockDependencies(String principal) {
-      when(mailService.isConfigured()).thenReturn(true);
-      when(configuration.getBaseUrl()).thenReturn("https://scm.hitchhiker.com");
-
-      when(mailTextResolver.getMailSubject(ENGLISH)).thenReturn("Awesome Subject");
-      when(mailTextResolver.getMailSubject(GERMAN)).thenReturn("Genialer Betreff");
-      when(mailTextResolver.getContentTemplatePath()).thenReturn("/path/to/template");
-      when(mailTextResolver.getContentTemplateModel(anyString(), anyBoolean())).then(ic -> {
-        boolean reviewer = ic.getArgument(1);
-        return reviewer ? reviewerModel : subscriberModel;
-      });
-
-      when(subject.getPrincipal()).thenReturn(principal);
-
-      when(envelopeBuilder
-        .withSubject("Awesome Subject")).thenReturn(subjectBuilder);
-      when(subjectBuilder
-        .withSubject(GERMAN, "Genialer Betreff")).thenReturn(subjectBuilder);
-      when(mailService.emailTemplateBuilder()).thenReturn(envelopeBuilder);
-    }
-
-    class Verifier {
-
-      Verifier send(String... users) {
-        for (String user : users) {
-          verify(envelopeBuilder).toUser(user);
-        }
-        return this;
-      }
-
-      Verifier notSend(String... users) {
-        for (String user : users) {
-          verify(envelopeBuilder, never()).toUser(user);
-        }
-        return this;
-      }
-
-      Verifier reviewer() {
-        verify(subjectBuilder.withTemplate("/path/to/template", MailTemplateType.MARKDOWN_HTML)).andModel(reviewerModel);
-        return this;
-      }
-
-      Verifier subscriber() {
-        verify(subjectBuilder.withTemplate("/path/to/template", MailTemplateType.MARKDOWN_HTML)).andModel(subscriberModel);
-        return this;
-      }
-    }
-
+    verify(envelopeBuilder).fromCurrentUser();
+    verify(envelopeBuilder).toUser("dent");
+    verify(envelopeBuilder).toUser("trillian");
   }
 
+  private void mockDependencies() {
+    when(mailService.isConfigured()).thenReturn(true);
+    when(configuration.getBaseUrl()).thenReturn("https://scm.hitchhiker.com");
+
+    when(mailTextResolver.getMailSubject(ENGLISH)).thenReturn("Awesome Subject");
+    when(mailTextResolver.getMailSubject(GERMAN)).thenReturn("Genialer Betreff");
+    when(mailTextResolver.getContentTemplatePath()).thenReturn("/path/to/template");
+    when(mailTextResolver.getContentTemplateModel(anyString())).thenReturn(subscriberModel);
+
+    when(envelopeBuilder
+      .withSubject("Awesome Subject")).thenReturn(subjectBuilder);
+    when(subjectBuilder
+      .withSubject(GERMAN, "Genialer Betreff")).thenReturn(subjectBuilder);
+    when(mailService.emailTemplateBuilder()).thenReturn(envelopeBuilder);
+  }
 
 }