+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Use mail topics so users can unsubscribe from mails for specific events ([#85](https://github.com/scm-manager/scm-review-plugin/pull/85))
+- Send mails for replies ([#85](https://github.com/scm-manager/scm-review-plugin/pull/85))
+
 ## [2.1.0] - 2020-06-19
 ### Added
 - Documentation in English and German ([#84](https://github.com/scm-manager/scm-review-plugin/pull/84))